remove old update scripts remove source code jquery & select2 upload last mysql structure

+++ docs/mysql/stat_table_User_stats.sql
@@ -10,6 +10,6 @@ CREATE TABLE `User_stats` (
   `router_id` int(11) DEFAULT 0,
   `auth_id` int(11) NOT NULL DEFAULT 0,
   `timestamp` datetime NOT NULL DEFAULT current_timestamp(),
-  `byte_in` int(11) NOT NULL DEFAULT 0,
-  `byte_out` int(11) NOT NULL DEFAULT 0
+  `byte_in` bigint(20) NOT NULL DEFAULT 0,
+  `byte_out` bigint(20) NOT NULL DEFAULT 0
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
